Properly handling multiple chunks on the same payload
Sometimes we receive more than one chunk on the same payload returned by
hackney. This causes some parts of the response from bedrock to be
missing and errors log on the format `{:bad_chunk, something, something}'.

This PR fixes that by extracting the rest of the original payload and
decoding it.